JBEHAVE-1241 Separated failing core stories.
