+++ examples/core/src/main/java/org/jbehave/examples/core/stories/failing/AfterScenarioOutcome.java
@@ -1,4 +1,4 @@
-package org.jbehave.examples.core.stories;
+package org.jbehave.examples.core.stories.failing;
 
 import org.jbehave.examples.core.CoreStory;
