+++ examples/core/src/main/java/org/jbehave/examples/core/stories/failing/SkipScenariosAfterFailure.java
@@ -1,4 +1,4 @@
-package org.jbehave.examples.core.stories;
+package org.jbehave.examples.core.stories.failing;
 
 import org.jbehave.core.configuration.Configuration;
 import org.jbehave.core.embedder.StoryControls;
