+++ examples/core/src/main/java/org/jbehave/examples/core/stories/failing/after_scenario_outcome.story
