+++ examples/core/src/main/java/org/jbehave/examples/core/stories/failing/failing_after_stories.story
