+++ examples/core/src/main/java/org/jbehave/examples/core/stories/failing/failure_correlation_one.story
