+++ examples/core/src/main/java/org/jbehave/examples/core/stories/failing/failure_followed_by_given_stories.story
