+++ examples/core/src/main/java/org/jbehave/examples/core/stories/failing/given_non_successful.story
