+++ examples/core/src/main/java/org/jbehave/examples/core/stories/failing/non_successful.story
