+++ examples/core/src/main/java/org/jbehave/examples/core/stories/failing/skip_scenarios_after_failure.story
