+++ examples/failing/src/main/java/org/jbehave/examples/core/FailingStories.java
@@ -1,0 +1,74 @@
+package org.jbehave.examples.core;
+
+import static org.jbehave.core.io.CodeLocations.codeLocationFromClass;
+import static org.jbehave.core.io.CodeLocations.codeLocationFromPath;
+import static org.jbehave.core.reporters.Format.CONSOLE;
+import static org.jbehave.core.reporters.Format.HTML_TEMPLATE;
+import static org.jbehave.core.reporters.Format.TXT;
+import static org.jbehave.core.reporters.Format.XML_TEMPLATE;
+
+import java.text.SimpleDateFormat;
+import java.util.List;
+import java.util.Properties;
+
+import org.jbehave.core.Embeddable;
+import org.jbehave.core.configuration.Configuration;
+import org.jbehave.core.configuration.MostUsefulConfiguration;
+import org.jbehave.core.context.Context;
+import org.jbehave.core.context.ContextView;
+import org.jbehave.core.context.JFrameContextView;
+import org.jbehave.core.embedder.PropertyBasedEmbedderControls;
+import org.jbehave.core.i18n.LocalizedKeywords;
+import org.jbehave.core.io.CodeLocations;
+import org.jbehave.core.io.LoadFromClasspath;
+import org.jbehave.core.io.StoryFinder;
+import org.jbehave.core.junit.JUnitStories;
+import org.jbehave.core.model.ExamplesTableFactory;
+import org.jbehave.core.model.TableTransformers;
+import org.jbehave.core.parsers.RegexPrefixCapturingPatternParser;
+import org.jbehave.core.parsers.RegexStoryParser;
+import org.jbehave.core.reporters.ContextOutput;
+import org.jbehave.core.reporters.CrossReference;
+import org.jbehave.core.reporters.Format;
+import org.jbehave.core.reporters.StoryReporterBuilder;
+import org.jbehave.core.steps.ContextStepMonitor;
+import org.jbehave.core.steps.InjectableStepsFactory;
+import org.jbehave.core.steps.InstanceStepsFactory;
+import org.jbehave.core.steps.ParameterConverters;
+import org.jbehave.core.steps.ParameterConverters.DateConverter;
+import org.jbehave.core.steps.ParameterConverters.ExamplesTableConverter;
+import org.jbehave.examples.core.service.TradingService;
+import org.jbehave.examples.core.steps.AndSteps;
+import org.jbehave.examples.core.steps.BeforeAfterSteps;
+import org.jbehave.examples.core.steps.CalendarSteps;
+import org.jbehave.examples.core.steps.CompositeSteps;
+import org.jbehave.examples.core.steps.ContextSteps;
+import org.jbehave.examples.core.steps.ExamplesTableParametersSteps;
+import org.jbehave.examples.core.steps.JsonSteps;
+import org.jbehave.examples.core.steps.MetaParametrisationSteps;
+import org.jbehave.examples.core.steps.MyContext;
+import org.jbehave.examples.core.steps.NamedParametersSteps;
+import org.jbehave.examples.core.steps.PendingSteps;
+import org.jbehave.examples.core.steps.PriorityMatchingSteps;
+import org.jbehave.examples.core.steps.RestartingSteps;
+import org.jbehave.examples.core.steps.SandpitSteps;
+import org.jbehave.examples.core.steps.SearchSteps;
+import org.jbehave.examples.core.steps.TableSteps;
+import org.jbehave.examples.core.steps.TraderSteps;
+
+/**
+ * <p>
+ * Example of how multiple stories can be run via JUnit.
+ * </p>
+ * <p>
+ * Stories are specified in classpath and correspondingly the
+ * {@link LoadFromClasspath} story loader is configured.
+ * </p>
+ */
+public class FailingStories extends CoreStories {
+
+    @Override
+    protected List<String> storyPaths() {
+        return new StoryFinder().findPaths(codeLocationFromClass(CoreStories.class), "**/failing/*.story","**/given_relative_path*");
+    }
+}
